requirements.txt
pandas
python-dotenv
duckdb
lxml
pyarrow
requests
pytest
pytest-mock
patch
plotly